Update index.html - paste real file
Because I hadn't validated my email address before putting this project on GitHub, it wasn't properly showing up on GitHub Pages, which was my ultimate goal.  Using Git Bash, I reset the index.html file and pushed it to github.com (that was my previous set of changes), and in this edit I'm pasting the actual index.html for this project into the editor.